drm: ttm/drm_vm: Add support for Marvell ThunderX2 to map memory differently
With ARM64 writecombine as NORMAL_NC we get suboptimal performance.
Since Marvell ThunderX2 support write combining with DEVICE_GRE mapping.
Apply the mapping change only if we are running on Marvell ThunderX2.

The performance improvement is huge with this.
A latency of 180ms is reduced to 2ms on a ffmpeg based encoding application.

Signed-off-by: George Cherian <george.cherian@marvell.com>